further optimise frame reading
Handle complete frames in one go, if possible, rather than header and
payload separately. This essentially halves the amount of binary
splitting in the framing code.

Note that we only do this when the buffer contains just one
binary. Tests have shown that attempting to introduce this
optimisation when the buffer comprises multiple binaries hurts
performance for large messages. That's presumably because we end up
constructing larger intermediate binaries.