Refactor the tests just a little bit more
Extract a helper to wait for process exits we're expecting. We ignore the
potential race (when establishing the monitor) since the exit reason is
irrelevant here, we just want to know the process is gone.

Instead of hitting 'nonode@nohost', which is slow (perhaps due to DNS
querying), we pull the current hostname and combine 'nonode' with that.