Improve synchronisation between plugins file and running broker
Since multiple brokers (on a host) may share the enabled-plugins file,
we may get out of sync. Diff any changes stemming from enable/disable
operations against the plugins actively running on the broker.

Extend the 'list' operation to display plugins that are [A] active
but not explicitly (or implicitly) enabled in the file and [I], those
that are explicitly (or implicitly) enabled but not actively running
in the broker.