ensure propagation of master death Previously slaves were monitoring the master and sending a process_death message through gm in an attempt to get the updated gm state propagated (and thus members_changed callbacks getting invoked on slaves, which in turn trigger promotion).
However, the DOWN notification for the master may well arrive and be
processed, and the process_death message sent around the ring, before
GM has noticed the master death, thus rendering the process_death
broadcast ineffectual for its intended purpose.

With the recent GM changes we can guarantee that at least one slave
will be notified of the master death via GM. So we emit the
process_death message then, and not on master DOWN (and ditch the
master monitoring as a result). Since the new emission is triggered by
a gm notification we know GM is aware of the master death, thus
avoiding the aforementioned problem.