record workers in a set ...instead of a queue.
That way when an idle worker is restarted (and sends an 'idle' message
to the pool), it won't end up in the pool twice.

Note that we always hand out work to the first worker in the
ordset. That is actually more efficient than the round-robin strategy
we had with the queue since it keeps a smaller number of workers busy
while others can hibernate.