monitor workers In a way this is a damage limitation exercise...
When a worker dies while working it is not in the pool anyway, so
noticing its death is a no-op.

When a worker dies while idle, it will not return to the pool when we
next submit work to it. But obviously the submitted work won't be
carried out, and the submitter will get an error (unless they
submitted the work asynchronously).

All the monitoring does is reduce the likelihood of the latter
happening. It cannot eliminate it though since the worker may die just
as work was being submitted to it.