Fix typo spotted by Edwin Fine.
Since hitting this line would cause a crash in the caller, and since this code is used very heavily (e.g. in the gen_server2 prioritised mailbox) and since it's been like that for over a year, I can only conclude that this code is unreachable. That conclusion is backed up by trying to get a pqueue() into a state which would hit that head - I was not able to. Still, it should be fixed...