Remove support for the legacy 'cluster_nodes' values
Before this change, a list of nodes without the node type was accepted.
In this case, the node type was guessed and a warning suggesting how to
update the configuration was logged.

Now, the node type is mandatory and the RabbitMQ server refuses to start
if the node type is unspecified.