Add plugin paths to the beginning of the code path
The goal is to make sure the right application is picked, in case the
same one is available as both a RabbitMQ plugin and an external Erlang
application. For instance, this could be the case with the Cowboy
application: a version is available in the plugins, but the user could
add an incompatible version to Erlang/OTP libdir or set ERL_LIBS to
point to it.

There's one exception currently: eldap. This application used to be
available as a 3rd party one. But since Erlang R15B01, it's included
in the standard library. We trust this version to have a stable API.
Therefore, if the node runs on R15B01 or later and eldap version is
1.0+, we use this one. In all other cases, we don't trust it and prefer
the RabbitMQ plugin.