Rearrange things a bit. We now have something that seems to work.
Merging a modified database back into a running cluster seems to be a
bit of a non-starter; Mnesia blows up with "Failed to merge schema:
Incompatible schema cookies." So instead we do something similar to if
we were doing upgrades: we always create the modified backup, but only
load it if we are the first node in the cluster to start; if not we
reset our mnesia state, resync with the cluster, then tell the cluster
to forget our old incarnation and update its queue records for the new
one.

This also has the advantage that we don't need two rename commands, we
just have the local offline one. Which is hopefully nicer
operationally.

It does mean that we go to the trouble of taking the backup and
modifying it only to throw it away if we turn out to be a secondary
node, but it should not be too expensive.