Add a new "keep_preferred" cluster partition handling method
The syntax is:
    {cluster_partition_management, {keep_preferred, node@domain}}

The specified node name is used to determine which partition should run
or be suspended. Nodes which can still reach the specified node continue
to run. Nodes which can't are suspended.

Compared to pause_minority, this allows the admin to determine which
nodes to prioritize in case of partitions with an equal number of nodes.