Throw an error if at least one plugin's module can't be loaded
This prevents a plugin from being enabled if it won't be able to
actually run later. A use case for this is a plugin built with Erlang
version N, but executed on Erlang version M, where M isn't capable of
running the bytecode from N. This was the case with Eralng R14B vs.
R15B.

The "rabbitmq-plugins enable <plugin>" command reports the error and the
plugin remains disabled.

A node reports the error too and refuses to start, exactly as if the
plugin was missing.