Autoheal: The loosing leader must wait for the winner_is message
As any other loosing nodes, the leader must wait for the winner_is
message, instead of restarting immediately.

The previous behaviour caused transient failures in the autoheal process
if the leader was in the middle of the restart at the time the winner
checks that all loosing nodes are up and running.