Rework SSL's verify_fun support
Until now, only the legacy form was supported (the one existing
in Erlang up-to R13B). The user would use the following RabbitMQ
configuration:
    {ssl_options, [
      {verify_fun, {Module, Function}}
    ]}

This is converted to the following ssl option:
    {ssl_options, [
      {verify_fun, fun(Errors) ->
                       Module:Function(Errors)
		   end}
    ]}

Although this form is still supported by Erlang R14B+, this is
undocumented and some users complain about RabbitMQ not matching the
expected behaviour, according to ssl's documentation.

Now, the new form is supported as well. Here's what the user would
configure:
    {ssl_options, [
      {verify_fun, {Module, Function, InitialUserState}}
    ]}
or:
    {ssl_options, [
      {verify_fun, {Module, Function}}
    ]}

This is converted to the new form:
    {ssl_options, [
      {verify_fun, {fun(OtpCert, Event, State) ->
                        Module:Function(OtpCert, Event, State)
		    end,
		    InitialUserState
		   }}
    ]}

To determine which form to use, we look at the version of the ssl
application:
    o  4.0.1+: We check if Module:Function/3 is exported and use the new
       form. If Module:Function/1 is exported, we use the old form. If
       both are exported, the new form is preferred. If InitialUserState
       is unspecified, 'none' is used.
    o  Before 4.0.1: We use the legacy form only.

If Module can't be loaded or if the expected Function/Arity is not
exported, an error is logged and RabbitMQ fails to start.