Autoheal: The leader waits for "done!" message from the winner
Before, the leader was monitoring the losers itself (exactly like the
winner). When they were all down, it was going back to the "not_healing"
state.

Therefore, there was a possibility that the leader and winner went
out-of-sync regarding the autoheal state.

Now, the leader simply waits for a confirmation from the winner that the
autoheal process is over. If the leader is a loser too, the autoheal
state is saved in the application environment to survive the restart.
When the leader is back up, it asks the winner to possibly notify it
again.