rabbitmq-env: Remove trailing slash from configurable paths
This fixes at least one problem with $RABBITMQ_PID_FILE which, by
default, is set to:
    RABBITMQ_PID_FILE=${RABBITMQ_MNESIA_DIR}.pid

In this case, if $RABBITMQ_MNESIA_DIR ends with a trailing slash, the
PID file is a hidden file placed in the Mnesia directory. This prevents
auto-clustering from working and this is tricky to understand why.

Fixes #118.