Make credit flow state transition interval 1s instead of 5s
Management UI and HTTP API currently report connections and channels
as in flow if they've been in flow for the last 5 seconds. That
can confuse the user with inter-node flow control, making them
believe the flow is permanent (it is not: the actual state
toggles many times a second).

This reduces the interval to 1s, which seems more reasonable and
accurate (in a way).

Fixes #138.