GM: Wait for messages to be ACK'd before exiting
When the GM is asked to stop, it flushes the pending messages but
does not ensure those messages are actually delivered to all members.

In some situations, this can lead to other GMs detecting that this GM is
unexpectedly down, even if this GM just sent a message to ask them to
stop as well.

When this happens in the case of queue mirroring for instance, this
causes a slave to promote itself as the new master, and the old master
to wait forever that other slaves exit.

Fixes #224.