rabbit_node_monitor: Rework `global` hang workaround
Up-to and including Erlang 17, `global` would use `erlang:now()` to tag
messages in the process dictionary. In Erlang 18.0, it uses a unique
integer, so we can't use this value to determine for how long messages
are stuck in the dictionary.

We now do two snapshots of the dictionary and compare them.

References #233.