Replace os:timestamp() with os:system_time(milli_seconds)
It will require all consumers of eg. events to be updated. But this is a
good occasion anyway to change the exposed timestamp format to something
pushed by the new Time API.

Comparisons and calculations are much simpler because the new format is
a simple integer.

The resolution is a millisecond, not a microsecond, because current
consumers of these timestamps do calculations in milliseconds.

References #233.