Priority queue: Fix a crash when querying 'head_message_timestamp'
Before the fix, the code would consider it an integer which could be
summed. It crashed if the head message had no timestamp at all (the
underlying queue returned '').

Now, we have a special case, exactly like 'backing_queue_status'. The
code first finds a priority with queued messages, then it queries the
underlying queue for the 'head_message_timestamp'.

Fixes #245.