improves push_betas_to_deltas performance
When messages are persistend to the index, instead of pushing one
message at a time to the FHC, now first we accumulate them in a list
that acts as a cache, and then we push them all at once once
push_betas_to_deltas finished recursion.

The same is done when marking the message as delivered.