Correct the description of hashing distribution
It's not true that "if you wish for queue A to receive twice as many messages as queue B then you bind the queue A with a binding key of twice the number [...] of the binding key of the binding to queue B."  The hashing is done on the routing keys, not the messages.  If routing key "foo" has 50,000 messages published and routing key "bar" has 1 message published, the ratio of messages received by queues A and B will not be 2:1.  In fact, all 50,001 messages might go to just one of those queues.

I've edited the langue to match the actual behavior and specifically clarified this confusion.