rabbitmq-plugins: Don't query rabbitmqctl if env vars are explicitly set
So if RABBITMQ_PLUGINS_DIR is explicitely set in the environment, do not
run rabbitmqctl. Likewise for RABBITMQ_ENABLED_PLUGINS_FILE.

This prevents rabbitmq-plugins from taking configuration from a running
unrelated node, when it is started from `make run-broker`. The make
command will still fail later but for the correct reason: there is
already a node conflicting with the one we want to start.

Without this, `make run-broker` would abort with the following error:
    Error: {cannot_write_enabled_plugins_file,
               "./sbin/../etc/rabbitmq/enabled_plugins",enoent}
... which doesn't make any sense.