rabbitmq-env: Revert "Use Erlang to determine local hostname"
After further investigation, the problem was not an inconsistency
between hostname(1) and Erlang, but a `/etc/rabbitmq/rabbitmq-env.conf`
in Travis CI's host having `NODENAME=rabbit@localhost`. See #402.

The real problem is therefore the testsuite which should ignore
`/etc/rabbitmq/rabbitmq-env.conf`.

The Windows side of the change is kept because it fixes a real issue.

References #402.