Fix memory reporting for connections
Before, there was two parts of the supervision tree: one for the
listener/acceptor, and one for connections. The latter was one
big supervisor handling all connections, regardless of the
listener they were coming from.

With Ranch, each listener has its own branch in the supervision
tree. Inside it are then two more branches, one for the
acceptors and one for the connections.

The fix simply adds all Ranch connection supervisors to the list
by fetching them from the ranch_server ETS table.