Fix backwards sort arrow icon.
It's conventional in GUIs to use a triangle or arrow icon to indicate how
data is currently sorted in a column - where the wide part of the triangle being
at the top and the narrow part at the bottom indicates that larger rows are above
(decreasing sort), and vice versa.  In other words, the direction in which the
icon width increases reflects the direction in which the values in the column increase.