make source-dist: Fix path to erlang.mk in rebar-based dependencies
When rebar is used to build a dependency, a Makefile including erlang.mk
is automatically generated. This Makefile finds erlang.mk in the project
which depends on this rebar-based component.

For instance, mochiweb could find erlang.mk in
rabbitmq_auth_backend_http. This breaks the build from a source
distribution if a project is not distributed (ie.
rabbitmq_auth_backend_http in this example).