Fix size check when marshalling lists/maps/arrays
Before, to determine if the small or large form should be used, the size
of the payload was checked, however, the space taken by the count of
elements was not accounted, even though it was later included in the
size. In pseudo-code:
    if ($size <= 255) then <store $size + 1 in the binary>

Therefore, when $size is exactly 255, ($size + 1) overflows and becomes
0.

Therefore, we now check for ($size + 1):
    if (($size + 1) <= 255) then <store $size + 1 in the binary>

While here, explicitely test the number of elements. It was not done
before as an optimization (because the number of elements can't be
greater than the size). But this renders the code less readable for no
real benefit. It should be easier to compare the test to the spec now.

Fixes #21.