Ensure queues are created with auto-delete and exclusive set to false
Since rabbitmq/rabbitmq-erlang-client#16, the `durable` parameter is
honored (instead of being forced to true). So AMQP 1.0 queues are
created as non-durable.

In our Erlang AMQP client, this means the queue is created with
auto-delete and exclusive set to true by default. We must explicitely
set those to false.

The testsuite is expanded to check that AMQP 0-9-1 queues created
out-of-band can be properly used from AMQP 1.0, no matter their
parameters.

Fixes #22.