rabbitmq-components.mk: Yet another attempt at getting the current Git ref
git-describe(1) always return the oldest matching tag, even if the
commit was checked out using a branch name or there is a newer tag. So
this command can't be used.

Now, we go back to parsing git-branch(1) output. It's kind of fragile
because the format is unstable (it changed at least once during Git
2.x). Furthermore it will fail if the branch name starts with a `(` and
contains `detached`.