Add support for hearbeats
This involves creating the keepalive supervisor when the connection
starts, and passing its pid onward to the handler. To achieve this
we put ourselves in just before starting cowboy_protocol, create
the supervisor and put its pid in the Cowboy environment. Later we
have a middleware that takes this pid and pass it to the MQTT handler
where it can then be used.

A few minor changes and fixes are also included in this commit,
including compacting the Req object and using hibernate.