rabbit_stomp_reader: Handle parsing errors gracefully
For a long time, process_received_bytes() didn't expect parsing
error, triggering a case_clause exception. That was ok before the merge
of #6 because go() caught the exception, logged an error and the process
terminated gracefully.

After #6, the exception was not caught anymore, causing a crash of the
process. Furthermore, terminate() was called with the case_clause
exception as the exit reason. It called log_network_error() which didn't
expect it either, causing an additional function_clause exception.

This double fault was noticed thanks to the
test_direct_client_connections_are_not_leaked() test which tries to open
100 faulty connections very quickly.

The broker couldn't handle such a rate in our CI, because of the
supervisor trying to cope with all those crashes, and the test failed
with a timeout.

So first, the parsing error is now handled in process_received_bytes()
case block. It triggers a normal termination of the process.

Then, the error is logged right there and the process stops with the
reason 'normal', because a non-'normal' error would still cause the
supervisor to log a message and handle recovery, which would still be
too slow.