"rabbitmqctl stop" exits with 0 if it fails to contact the node
If rabbitmqctl fails to contact the node to ask it to stop, it still
displays an error message. However, now it exits with 0; it was
EX_UNAVAILABLE before.

There are two main reasons it can happen:
    o  The node is stopped, so that's ok to exit with 0 because the end
       result is the same.
    o  The Erlang cookie is incorrect. The error message is here to help
       but this situation is quite rare compared to the other one.

There is no way to distinguish both, so let's assume the common
situation.

This helps in the case of systemd in particular: if one stops RabbitMQ
using "rabbitmqctl stop" instead of "service rabbitmq-server stop",
systemd tries to call "rabbitmqctl stop" again. Before this change, it
failed and put the service in a "failed" state.

Fixes #693.