Allows watermark to be specified as an absolute value
Rabbit allows the watermark to be specified as an absolute value, or an
absolute value with memory units .. such as {:absolute, N} which can be
specified as 'absolute 10' or 'absolute 10MB'.  In the case of the
latter, the value in bytes is sent as N.

This commit adds some useful functions for calculating the absolute
values to the Helpers module.