make docs: Force a dependency to `edoc`
In newer Erlang.mk, `edoc` is added as a dependency to `docs` only if
`doc/overview.edoc` exists. However in our Makefile, this file is
generated. Therefore, at the time the Makefile is evaluated, this file
is missing and `edoc` is not added as a dependency to `docs`.

To workaround this, we explicitely add this dependency.