rabbit_ct_helpers: Use arbitrary node name and TCP ports
Furthermore, if the node fails to start, try again with another name and
set of TCP ports until we find a setup which boots fine.

This allows to run the testsuite while there is another unrelated
RabbitMQ node already running. Moreover, this avoids unrelated AMQP
clients to mess with the testsuite node.