Improve notifications about damaged I/O subsystem
Follow-up to https://github.com/rabbitmq/rabbitmq-server/pull/747

It's the best thing that can be done. Without `exit` anything inside
rabbit could hang - most prone to it is the `start_app` action. Crashing
loudly will be helpful, because during 2016 there were at least 2
production investigations about that issue.

And here is the script that I was using for reproduction:
https://gist.github.com/binarin/edd4102b7262d53e2cf09c76d5ca8e9e