rabbitmq-run.mk: Do not use ?= in target-specific variable assignments
This appears to be broken with at least GNU Make 3.81 on Debian Wheezy,
when the variable is overriden: instead of the new value, the variable
contains garbage.

This can be demonstrated with the following Makefile:
    TEST_CONFIG_FILE = <default>

    all: RABBITMQ_CONFIG_FILE ?= $(TEST_CONFIG_FILE)
    all:
            @echo TEST_CONFIG_FILE="$(TEST_CONFIG_FILE)"
            @echo RABBITMQ_CONFIG_FILE="$(RABBITMQ_CONFIG_FILE)"

The expected output of `make RABBITMQ_CONFIG_FILE=/tmp` is:
    TEST_CONFIG_FILE=<default>
    RABBITMQ_CONFIG_FILE=/tmp

But GNU Make 3.81 on Debian Wheezy produces:
    TEST_CONFIG_FILE=<default>
    RABBITMQ_CONFIG_FILE= �?

In the end, using ?= is not required here, a plain = is enough: the
problem is worked around and the caller can still override the variable.

Fixes #93.