config_schema_SUITE: Copy Cuttlefish schema in the testsuite data directory
We must do that instead of adding the plugins as test dependencies
because doing this introduces a circular dependency: all plugins depend
on rabbitmq-server.

With the circular dependency in place, the broker was compiled with
$(ERLC_OPTS) compiler flags, because the build was triggered by the
first plugin. Therefore we missed all the $(TEST_ERLC_OPTS) additions.

With the circular dependency fixed, the broker is again built with
$(TEST_ERLC_OPTIS): `src/truncate.erl` depends on the `-DTEST=1` flag.

In the longer term, we must move the testcases depending on the plugins'
schemas to their corresponding plugin.